% Auto-generated: do not edit by hand
\name{dmeMdxEditor}

\alias{dmeMdxEditor}

\title{MdxEditor component}

\description{
MdxEditor is a markdown editor that allows to preview the markdown content and edit as rich text.
}

\usage{
dmeMdxEditor(id=NULL, default_text=NULL, text=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{default_text}{Character. The initial text of the editor}

\item{text}{Character. The text of the editor (read)}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

